# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "vLLM CI Dashboard API",
    "version": "1.0.0",
    "description": "Public read-only API backing https://ci.vllm.ai. All documented GET endpoints are unauthenticated and return JSON; errors use `{ \"error\": string }`. Responses are cached via `Vercel-CDN-Cache-Control`; treat data as eventually consistent (30s-10min staleness)."
  },
  "servers": [
    {
      "url": "https://ci.vllm.ai",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/builds/summary": {
      "get": {
        "operationId": "get-builds-summary",
        "summary": "Compact build summary for agents and CLI use",
        "description": "Defaults to a text/plain summary; pass format=json for JSON.",
        "tags": [
          "Builds"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pipeline",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "branch",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "number",
              "description": "0-based (default 0)"
            },
            "required": false,
            "description": "0-based (default 0)"
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "type": "number",
              "description": "clamped to 1-30 (default 10)"
            },
            "required": false,
            "description": "clamped to 1-30 (default 10)"
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "text",
                "json"
              ],
              "description": "default \"text\" returns text/plain"
            },
            "required": false,
            "description": "default \"text\" returns text/plain"
          },
          {
            "in": "query",
            "name": "jobs",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ],
              "description": "\"false\" omits failed job detail"
            },
            "required": false,
            "description": "\"false\" omits failed job detail"
          },
          {
            "in": "query",
            "name": "jobGroups",
            "schema": {
              "type": "string",
              "description": "comma-separated group names"
            },
            "required": false,
            "description": "comma-separated group names"
          },
          {
            "in": "query",
            "name": "jobNames",
            "schema": {
              "type": "string",
              "description": "comma-separated exact job names"
            },
            "required": false,
            "description": "comma-separated exact job names"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "operationId": "get-jobs",
        "summary": "Job failure ranking and duration stats",
        "description": "Without startDate/endDate, only jobs with runs in the last 7 days are included.",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pipeline",
            "schema": {
              "type": "string",
              "description": "default \"CI\""
            },
            "required": false,
            "description": "default \"CI\""
          },
          {
            "in": "query",
            "name": "branch",
            "schema": {
              "type": "string",
              "description": "default \"main\""
            },
            "required": false,
            "description": "default \"main\""
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
...
}